LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbobjectstore_add5-expected.txt


PASS IDBObjectStore.add() - object store's key path is an object attribute
